Updated sellers model
